requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.6
django-allauth==0.45.0
django-bootstrap-modal-forms==2.2.0
django-cleanup==5.2.0
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.12.4
environ==1.0
gunicorn==20.1.0
idna==3.2
instance==0.0.0.0.0.0a0.dev20210120
lark-parser==0.11.3
lxml==4.6.3
numpy==1.21.1
oauthlib==3.1.1
pandas==1.3.1
phonenumbers==8.12.29
Pillow==8.3.1
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
required==0.4.0
sentry-sdk==1.3.1
simplejson==3.17.5
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
Unidecode==1.2.0
urllib3==1.26.6
whitenoise==5.3.0
